Fix the issue where Windows containers could fail with a NumberFormatException.
Windows containers sometimes exit with 32-bit unsigned exit codes, but
we only supported 32-bit signed values.

+++ app/src/main/kotlin/batect/docker/DockerTypes.kt
@@ -32,7 +32,7 @@ data class DockerVolumeMount(val localPath: String, val containerPath: String, v
     override fun toString(): String = if (options == null) "$localPath:$containerPath" else "$localPath:$containerPath:$options"
 }
 
-data class DockerContainerRunResult(val exitCode: Int)
+data class DockerContainerRunResult(val exitCode: Long)
 
 @Serializable
 data class DockerNetwork(val id: String)
