+++ app/src/main/kotlin/batect/ui/quiet/QuietEventLogger.kt
@@ -39,6 +39,6 @@ class QuietEventLogger(
 
     override fun onTaskFailed(taskName: String, manualCleanupInstructions: TextRun) {}
     override fun onTaskStarting(taskName: String) {}
-    override fun onTaskFinished(taskName: String, exitCode: Int, duration: Duration) {}
+    override fun onTaskFinished(taskName: String, exitCode: Long, duration: Duration) {}
     override fun onTaskFinishedWithCleanupDisabled(manualCleanupInstructions: TextRun) {}
 }
