+++ app/src/main/kotlin/batect/ui/simple/SimpleEventLogger.kt
@@ -190,7 +190,7 @@ class SimpleEventLogger(
         console.println(Text.white(Text("Running ") + Text.bold(taskName) + Text("...")))
     }
 
-    override fun onTaskFinished(taskName: String, exitCode: Int, duration: Duration) {
+    override fun onTaskFinished(taskName: String, exitCode: Long, duration: Duration) {
         console.println(Text.white(Text.bold(taskName) + Text(" finished with exit code $exitCode in ${duration.humanise()}.")))
     }
